+++ sched/mqueue/mq_msgfree.c
@@ -31,6 +31,7 @@
 #include <nuttx/irq.h>
 #include <nuttx/arch.h>
 #include <nuttx/kmalloc.h>
+#include <nuttx/spinlock.h>
 
 #include "mqueue/mqueue.h"
 
@@ -56,6 +57,8 @@
 
 void nxmq_free_msg(FAR struct mqueue_msg_s *mqmsg)
 {
+  irqstate_t flags;
+
   /* If this is a generally available pre-allocated message,
    * then just put it back in the free list.
    */
@@ -66,7 +69,9 @@ void nxmq_free_msg(FAR struct mqueue_msg_s *mqmsg)
        * list from interrupt handlers.
        */
 
+      flags = spin_lock_irqsave(NULL);
       list_add_tail(&g_msgfree, &mqmsg->node);
+      spin_unlock_irqrestore(NULL, flags);
     }
 
   /* If this is a message pre-allocated for interrupts,
@@ -79,7 +84,9 @@ void nxmq_free_msg(FAR struct mqueue_msg_s *mqmsg)
        * list from interrupt handlers.
        */
 
+      flags = spin_lock_irqsave(NULL);
       list_add_tail(&g_msgfreeirq, &mqmsg->node);
+      spin_unlock_irqrestore(NULL, flags);
     }
 
   /* Otherwise, deallocate it.  Note:  interrupt handlers
