+++ sched/mqueue/mq_sndinternal.c
@@ -37,7 +37,6 @@
 #include <debug.h>
 
 #include <nuttx/irq.h>
-#include <nuttx/kmalloc.h>
 #include <nuttx/arch.h>
 #include <nuttx/sched.h>
 #include <nuttx/cancelpt.h>
@@ -46,143 +45,56 @@
 #include "mqueue/mqueue.h"
 
 /****************************************************************************
- * Public Functions
+ * Private Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: nxmq_verify_send
+ * Name: nxmq_sndtimeout
  *
  * Description:
- *   This is internal, common logic shared by both [nx]mq_send and
- *   [nx]mq_timesend.  This function verifies the input parameters that are
- *   common to both functions.
+ *   This function is called if the timeout elapses before the message queue
+ *   becomes non-full.
  *
  * Input Parameters:
- *   msgq   - Message queue descriptor
- *   msg    - Message to send
- *   msglen - The length of the message in bytes
- *   prio   - The priority of the message
+ *   arg - The argument that was provided when the timeout was configured.
  *
  * Returned Value:
- *   On success, 0 (OK) is returned. On failure, a negated errno value is
- *   returned.
- *
- *     EINVAL   Either msg or msgq is NULL or the value of prio is invalid.
- *     EBADF    Message queue opened not opened for writing.
- *     EMSGSIZE 'msglen' was greater than the maxmsgsize attribute of the
- *               message queue.
- *
- ****************************************************************************/
-
-#ifdef CONFIG_DEBUG_FEATURES
-int nxmq_verify_send(FAR struct file *mq, FAR const char *msg,
-                     size_t msglen, unsigned int prio)
-{
-  FAR struct inode *inode = mq->f_inode;
-  FAR struct mqueue_inode_s *msgq;
-
-  if (inode == NULL)
-    {
-      return -EBADF;
-    }
-
-  msgq = inode->i_private;
-
-  /* Verify the input parameters */
-
-  if (msg == NULL || msgq == NULL || prio >= MQ_PRIO_MAX)
-    {
-      return -EINVAL;
-    }
-
-  if ((mq->f_oflags & O_WROK) == 0)
-    {
-      return -EBADF;
-    }
-
-  if (msglen > (size_t)msgq->maxmsgsize)
-    {
-      return -EMSGSIZE;
-    }
-
-  return OK;
-}
-#endif
-
-/****************************************************************************
- * Name: nxmq_alloc_msg
- *
- * Description:
- *   The nxmq_alloc_msg function will get a free message for use by the
- *   operating system.  The message will be allocated from the g_msgfree
- *   list.
- *
- *   If the list is empty AND the message is NOT being allocated from the
- *   interrupt level, then the message will be allocated.  If a message
- *   cannot be obtained, the operating system is dead and therefore cannot
- *   continue.
- *
- *   If the list is empty AND the message IS being allocated from the
- *   interrupt level.  This function will attempt to get a message from
- *   the g_msgfreeirq list.  If this is unsuccessful, the calling interrupt
- *   handler will be notified.
- *
- * Input Parameters:
  *   None
  *
- * Returned Value:
- *   A reference to the allocated msg structure.  On a failure to allocate,
- *   this function PANICs.
+ * Assumptions:
  *
  ****************************************************************************/
 
-FAR struct mqueue_msg_s *nxmq_alloc_msg(void)
+static void nxmq_sndtimeout(wdparm_t arg)
 {
-  FAR struct list_node *mqmsg;
+  FAR struct tcb_s *wtcb = (FAR struct tcb_s *)(uintptr_t)arg;
+  irqstate_t flags;
 
-  /* Try to get the message from the generally available free list. */
+  /* Disable interrupts.  This is necessary because an interrupt handler may
+   * attempt to send a message while we are doing this.
+   */
 
-  mqmsg = list_remove_head(&g_msgfree);
-  if (mqmsg == NULL)
+  flags = enter_critical_section();
+
+  /* It is also possible that an interrupt/context switch beat us to the
+   * punch and already changed the task's state.
+   */
+
+  if (wtcb->task_state == TSTATE_WAIT_MQNOTFULL)
     {
-      /* If we were called from an interrupt handler, then try to get the
-       * message from generally available list of messages. If this fails,
-       * then try the list of messages reserved for interrupt handlers
-       */
+      /* Restart with task with a timeout error */
 
-      if (up_interrupt_context())
-        {
-          /* Try the free list reserved for interrupt handlers */
-
-          mqmsg = list_remove_head(&g_msgfreeirq);
-        }
-
-      /* We were not called from an interrupt handler. */
-
-      else
-        {
-          /* If we cannot a message from the free list, then we will have to
-           * allocate one.
-           */
-
-          mqmsg = (FAR struct list_node *)
-            kmm_malloc((sizeof (struct mqueue_msg_s)));
-
-          /* Check if we allocated the message */
-
-          if (mqmsg != NULL)
-            {
-              /* Yes... remember that this message was dynamically
-               * allocated.
-               */
-
-              ((FAR struct mqueue_msg_s *)mqmsg)->type = MQ_ALLOC_DYN;
-            }
-        }
+      nxmq_wait_irq(wtcb, ETIMEDOUT);
     }
 
-  return (FAR struct mqueue_msg_s *)mqmsg;
+  /* Interrupts may now be re-enabled. */
+
+  leave_critical_section(flags);
 }
+
+/****************************************************************************
+ * Public Functions
+ ****************************************************************************/
 
 /****************************************************************************
  * Name: nxmq_wait_send
@@ -194,7 +106,7 @@ FAR struct mqueue_msg_s *nxmq_alloc_msg(void)
  *
  * Input Parameters:
  *   msgq   - Message queue descriptor
- *   oflags - flags from user set
+ *   abstime - The absolute time to wait until
  *
  * Returned Value:
  *   On success, nxmq_wait_send() returns 0 (OK); a negated errno value is
@@ -212,9 +124,11 @@ FAR struct mqueue_msg_s *nxmq_alloc_msg(void)
  *
  ****************************************************************************/
 
-int nxmq_wait_send(FAR struct mqueue_inode_s *msgq, int oflags)
+int nxmq_wait_send(FAR struct mqueue_inode_s *msgq,
+                   FAR const struct timespec *abstime,
+                   sclock_t ticks)
 {
-  FAR struct tcb_s *rtcb;
+  FAR struct tcb_s *rtcb = this_task();
 
 #ifdef CONFIG_CANCELLATION_POINTS
   /* nxmq_wait_send() is not a cancellation point, but may be called via
@@ -231,6 +145,17 @@ int nxmq_wait_send(FAR struct mqueue_inode_s *msgq, int oflags)
     }
 #endif
 
+  if (abstime)
+    {
+      wd_start_realtime(&rtcb->waitdog, abstime,
+                        nxmq_sndtimeout, (wdparm_t)rtcb);
+    }
+  else if (ticks >= 0)
+    {
+      wd_start(&rtcb->waitdog, ticks,
+               nxmq_sndtimeout, (wdparm_t)rtcb);
+    }
+
   /* Verify that the queue is indeed full as the caller thinks */
 
   /* Loop until there are fewer than max allowable messages in the
@@ -239,17 +164,6 @@ int nxmq_wait_send(FAR struct mqueue_inode_s *msgq, int oflags)
 
   while (msgq->nmsgs >= msgq->maxmsgs)
     {
-      /* Should we block until there is sufficient space in the
-       * message queue?
-       */
-
-      if ((oflags & O_NONBLOCK) != 0)
-        {
-          /* No... We will return an error to the caller. */
-
-          return -EAGAIN;
-        }
-
       /* Block until the message queue is no longer full.
        * When we are unblocked, we will try again
        */
@@ -291,85 +205,40 @@ int nxmq_wait_send(FAR struct mqueue_inode_s *msgq, int oflags)
 
       if (rtcb->errcode != OK)
         {
-          return -rtcb->errcode;
+          break;
         }
     }
 
-  return OK;
+  if (abstime || ticks >= 0)
+    {
+      wd_cancel(&rtcb->waitdog);
+    }
+
+  return -rtcb->errcode;
 }
 
 /****************************************************************************
- * Name: nxmq_do_send
+ * Name: nxmq_notify_send
  *
  * Description:
- *   This is internal, common logic shared by both [nx]mq_send and
- *   [nx]mq_timesend.  This function adds the specified message (msg) to the
- *   message queue (msgq).  Then it notifies any tasks that were waiting
- *   for message queue notifications setup by mq_notify.  And, finally, it
- *   awakens any tasks that were waiting for the message not empty event.
+ *   This function is called when a message is sent to a message queue.
+ *   It will notify any tasks that are waiting for the message queue to be
+ *   non-full.
  *
  * Input Parameters:
  *   msgq   - Message queue descriptor
- *   msg    - Message to send
- *   msglen - The length of the message in bytes
- *   prio   - The priority of the message
  *
  * Returned Value:
- *   This function always returns OK.
+ *   None
+ *
+ * Assumptions/restrictions:
+ * - Executes within a critical section established by the caller.
  *
  ****************************************************************************/
 
-int nxmq_do_send(FAR struct mqueue_inode_s *msgq,
-                 FAR struct mqueue_msg_s *mqmsg,
-                 FAR const char *msg, size_t msglen, unsigned int prio)
+void nxmq_notify_send(FAR struct mqueue_inode_s *msgq)
 {
-  FAR struct mqueue_msg_s *prev = NULL;
-  FAR struct mqueue_msg_s *next;
   FAR struct tcb_s *btcb;
-
-  /* Construct the message header info */
-
-  mqmsg->priority = prio;
-  mqmsg->msglen   = msglen;
-
-  /* Copy the message data into the message */
-
-  memcpy((FAR void *)mqmsg->mail, (FAR const void *)msg, msglen);
-
-  /* Insert the new message in the message queue
-   * Search the message list to find the location to insert the new
-   * message. Each is list is maintained in ascending priority order.
-   */
-
-  list_for_every_entry(&msgq->msglist, next, struct mqueue_msg_s, node)
-    {
-      if (prio > next->priority)
-        {
-          break;
-        }
-      else
-        {
-          prev = next;
-        }
-    }
-
-  /* Add the message at the right place */
-
-  if (prev)
-    {
-      list_add_after(&prev->node, &mqmsg->node);
-    }
-  else
-    {
-      list_add_head(&msgq->msglist, &mqmsg->node);
-    }
-
-  /* Increment the count of messages in the queue */
-
-  if (msgq->nmsgs++ == 0)
-    {
-      nxmq_pollnotify(msgq, POLLIN);
-    }
 
   /* Check if we need to notify any tasks that are attached to the
    * message queue
@@ -437,6 +306,4 @@ int nxmq_do_send(FAR struct mqueue_inode_s *msgq,
           up_switch_context(btcb, rtcb);
         }
     }
-
-  return OK;
 }
