+++ sched/mqueue/mqueue.h
@@ -109,6 +109,9 @@ struct task_group_s; /* Forward reference */
 /* Functions defined in mq_initialize.c *************************************/
 
 void nxmq_initialize(void);
+
+/* mq_msgfree.c *************************************************************/
+
 void nxmq_free_msg(FAR struct mqueue_msg_s *mqmsg);
 
 /* mq_waitirq.c *************************************************************/
@@ -117,30 +120,18 @@ void nxmq_wait_irq(FAR struct tcb_s *wtcb, int errcode);
 
 /* mq_rcvinternal.c *********************************************************/
 
-#ifdef CONFIG_DEBUG_FEATURES
-int nxmq_verify_receive(FAR struct file *mq, FAR char *msg, size_t msglen);
-#else
-#  define nxmq_verify_receive(msgq, msg, msglen) OK
-#endif
 int nxmq_wait_receive(FAR struct mqueue_inode_s *msgq,
-                      int oflags, FAR struct mqueue_msg_s **rcvmsg);
-ssize_t nxmq_do_receive(FAR struct mqueue_inode_s *msgq,
-                        FAR struct mqueue_msg_s *mqmsg,
-                        FAR char *ubuffer, FAR unsigned int *prio);
+                      FAR struct mqueue_msg_s **rcvmsg,
+                      FAR const struct timespec *abstime,
+                      sclock_t ticks);
+void nxmq_notify_receive(FAR struct mqueue_inode_s *msgq);
 
 /* mq_sndinternal.c *********************************************************/
 
-#ifdef CONFIG_DEBUG_FEATURES
-int nxmq_verify_send(FAR struct file *mq, FAR const char *msg,
-                     size_t msglen, unsigned int prio);
-#else
-#  define nxmq_verify_send(mq, msg, msglen, prio) OK
-#endif
-FAR struct mqueue_msg_s *nxmq_alloc_msg(void);
-int nxmq_wait_send(FAR struct mqueue_inode_s *msgq, int oflags);
-int nxmq_do_send(FAR struct mqueue_inode_s *msgq,
-                 FAR struct mqueue_msg_s *mqmsg,
-                 FAR const char *msg, size_t msglen, unsigned int prio);
+int nxmq_wait_send(FAR struct mqueue_inode_s *msgq,
+                   FAR const struct timespec *abstime,
+                   sclock_t ticks);
+void nxmq_notify_send(FAR struct mqueue_inode_s *msgq);
 
 /* mq_recover.c *************************************************************/
